refactorlogic
